+++ src/SecucardConnect/SecucardConnect.php
@@ -33,7 +33,7 @@ final class SecucardConnect
     /**
      * SDK version
      */
-    const VERSION = '1.10.1';
+    const VERSION = '1.11.0';
 
     /**
      * @var OAuthProvider