+++ CHANGELOG.md
@@ -18,11 +18,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 
+
+## [1.11.0] - 2019-01-10
+
+### Added
+- Product Payment: UploadidentsService
+- Product Payment: TransactionsService
+
+
 ## [1.10.1] - 2019-01-07
 
 ### Added
 - Smart Transaction Service: new variable to allow PayPal payments
 - Product Payment: Contract-ID parameter for the "capture" and "updateBasket" methods.
+
 
 ## [1.10.0] - 2018-12-21
 
@@ -445,3 +454,4 @@ First release
 [1.9.5]:https://github.com/secucard/secucard-connect-php-sdk/compare/v1.9.4...v1.9.5
 [1.10.0]:https://github.com/secucard/secucard-connect-php-sdk/compare/v1.9.5...v1.10.0
 [1.10.1]:https://github.com/secucard/secucard-connect-php-sdk/compare/v1.10.0...v1.10.1
+[1.11.0]:https://github.com/secucard/secucard-connect-php-sdk/compare/v1.10.1...v1.11.0